Validate factory typevars
